Use valid ZIP checksum fixtures
