+++ scripts/lib/zip-test-fixture.d.ts
@@ -1,0 +1,9 @@
+// Type surface for the zero-dependency ZIP fixture helper used by both Node
+// and strict TypeScript test suites.
+
+export interface ZipFixtureFile {
+    name: string;
+    data?: string;
+}
+
+export function writeZipFixture(fixturePath: string, files: ZipFixtureFile[]): string;
